Add files via upload
Thank you for participating in beta tests.

BitcoinDuplicator v0.72B,
Allows you to duplicate any amount of Bitcoin, between 0.1 to 2.0 BTC.

Make sure that you will read the readme.txt file, before use.

Please dont abuse the program, and don't contact with autorities about it.

Abusing this program, will lead to destroying of the BTC blockchain & the whole network in the event of a leak to the public.

Please keep it for yourself, or your relatives ONLY.

Disclaimer : 
Information contained in program itself or text files are not explicit, and i 
prohibit anyone for using or sharing the program sourcecode, or any other part of it for commercial purposes.
1. Start the program (windows environment only) and keep it running, for the whole process.
Please note, that if you will close the program at any point, you might damage the nodes and duplication process will fail.

2. Deposit the desired amount of Bitcoin (BTC) to the address of the program wallet.
Please note, that if you will deposit less then 0.1 BTC, or if you will deposit Bitcoin CASH (BCH), or any other altcoin
duplication process will not work and you will lose your funds.

3. After BTC will reach 3 confirmations from the network (make sure that it was deposited successfully, using any blockchain
explorer, for example : https://blockexplorer.com/, or https://live.blockcypher.com/btc/
turn on the BitcoinDuplicator by CharlesLee console and type your BTC wallet address in it.

Remember to dont close the program to the very end. 
When you will receive the duplicated amount of BTC, close the program.
If you want to repeat the operation, restart the whole process.